+++ src/commands/slashCommands/map/map.js
@@ -27,12 +27,7 @@ module.exports = {
         var gamemode = interaction.options.get('gamemode')?.value ?? "battle_royale";
         var url = encodeURI(`https://api.mozambiquehe.re/maprotation?version=2&auth=${auth}`);
         fetch(url)
-            .then(res => {
-                if (res.status === 200) { return res.json() } else {
-                    handleError(interaction, userData, res.status);
-                    return Promise.reject('Error occurred');
-                }
-            })
+            .then(res => res.status === 200 ? res.json() : handleError(interaction, userData, res.status))
             .then(data => {
                 var mapDescr = getMapDescription(gamemode, data, langOpt);
                 const mapEmbed = new EmbedBuilder()