+++ src/commands/slashCommands/user/me.js
@@ -2,7 +2,7 @@ const { SlashCommandBuilder, EmbedBuilder, InteractionContextType } = require('d
 const fetch = require('node-fetch');
 const lang = require('../../../data/lang/lang.json');
 const emoji = require('../../../data/utilities/emoji.json');
-const { getStatus, handleError } = require('../../../utilities/functions/utilities');
+const { getStatus, handleError, checkData } = require('../../../utilities/functions/utilities');
 require('dotenv').config();
 
 module.exports = {
@@ -24,13 +24,10 @@ module.exports = {
 
             var url = encodeURI(`https://api.mozambiquehe.re/bridge?version=5&platform=${platform}&player=${player}&auth=${auth}`);
             fetch(url)
-                .then(res => {
-                    if (res.status === 200) { return res.json() } else {
-                        handleError(interaction, userData, res.status);
-                        return Promise.reject('Error occurred');
-                    }
-                })
+                .then(res => res.status === 200 ? res.json() : handleError(interaction, userData, res.status))
                 .then(data => {
+                    checkData(data, interaction, userData);
+
                     var badge1 = data?.legends?.selected?.data?.[0];
                     var badge2 = data?.legends?.selected?.data?.[1];
                     var badge3 = data?.legends?.selected?.data?.[2];