+++ src/commands/slashCommands/utilities/crafting.js
@@ -18,12 +18,7 @@ module.exports = {
 
         var url = encodeURI(`https://api.mozambiquehe.re/crafting?auth=${auth}`);
         fetch(url)
-            .then(res => {
-                if (res.status === 200) { return res.json() } else {
-                    handleError(interaction, userData, res.status);
-                    return Promise.reject('Error occurred');
-                }
-            })
+            .then(res => res.status === 200 ? res.json() : handleError(interaction, userData, res.status))
             .then(async data => {
                 //top lef
                 const daily1 = await Canvas.loadImage(data[4].bundleContent[0].itemType.asset);