+++ src/commands/slashCommands/utilities/news.js
@@ -14,12 +14,7 @@ module.exports = {
 
         var url = `https://api.mozambiquehe.re/news?auth=${auth}`;
         fetch(url)
-            .then(res => {
-                if (res.status === 200) { return res.json() } else {
-                    handleError(interaction, langOpt, res.status);
-                    return Promise.reject('Error occurred');
-                }
-            })
+            .then(res => res.status === 200 ? res.json() : handleError(interaction, userData, res.status))
             .then(async data => {
 
                 const frist = new ButtonBuilder()