+++ src/commands/subcommands/embedColor.js
@@ -16,7 +16,8 @@ const embedColorSubComand = (subCommand) => subCommand
             .setDescriptionLocalizations({ nl: 'maak een keuze.' })
             .setRequired(true)
             .addChoices(
-                { name: 'Default', value: embedColor },
+                { name: 'Default', value: 'Default' },
+                { name: 'Invisible', value: embedColor },
                 { name: "White", value: "White" },
                 { name: 'Yellow', value: 'Yellow' },
                 { name: 'Orange', value: 'orange' },
@@ -43,30 +44,23 @@ const embedColorSubFunction = async (interaction, auth, userData) => {
         .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
         .setTimestamp()
 
-    var option = interaction.options.get('option').value;
-    var customColor = interaction.options?.getString('color');
-    if (customColor && !customColor.startsWith('#')) customColor = '#' + customColor;
+    var errorEmbed = new EmbedBuilder()
+        .setTitle('ERROR')
+        .setDescription(lang[langOpt].settings.line_10)
+        .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
+        .setColor('Red')
+        .setTimestamp()
 
+    let option = interaction.options.get('option').value;
+    const customColor = interaction.options?.getString('color')?.startsWith('#') ? interaction.options.getString('color') : `#${interaction.options.getString('color')}`;
     if (option === "Custom") {
-        if (isHexColor(customColor) == true) {
-            await setDoc(doc(db, 'users', interaction.user.id), { embedColor: customColor }, { merge: true });
-            colorEmbed.setDescription(lang[langOpt].settings.line_3 + "``" + customColor + "``.")
-                .setColor(customColor);
-        } else {
-            var errorEmbed = new EmbedBuilder()
-                .setTitle('ERROR')
-                .setDescription(lang[langOpt].settings.line_10)
-                .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
-                .setColor('Red')
-                .setTimestamp()
-
-            return interaction.editReply({ embeds: [errorEmbed], ephemeral: userData.invisible });
-        }
-    } else {
-        await setDoc(doc(db, 'users', interaction.user.id), { embedColor: option }, { merge: true });
-        colorEmbed.setDescription(lang[langOpt].settings.line_3 + "``" + option + "``.")
-            .setColor(option);
+        if (isHexColor(customColor) == false) return interaction.editReply({ embeds: [errorEmbed], ephemeral: userData.invisible });
+        option = customColor;
     }
+
+    await setDoc(doc(db, 'users', interaction.user.id), { embedColor: option }, { merge: true });
+    colorEmbed.setDescription(lang[langOpt].settings.line_3 + "``" + option + "``.")
+        .setColor(option);
 
     interaction.editReply({ embeds: [colorEmbed], ephemeral: userData.invisible });
 }