+++ src/commands/subcommands/link.js
@@ -4,7 +4,7 @@ const firebase = require('firebase/app');
 const { getFirestore, doc, setDoc } = require('firebase/firestore');
 const firebaseConfig = require('../../SECURITY/firebaseConfig.json');
 const lang = require('../../data/lang/lang.json');
-const { handleError } = require('../../utilities/functions/utilities');
+const { handleError, checkData } = require('../../utilities/functions/utilities');
 const app = firebase.initializeApp(firebaseConfig);
 const db = getFirestore(app);
 
@@ -33,49 +33,43 @@ const linkSubFunction = async (interaction, auth, userData) => {
     var langOpt = userData.lang;
     await interaction.deferReply({ ephemeral: userData.invisible });
 
-    var platform = interaction.options.get('platform').value
+    var platform = interaction.options.get('platform').value;
     const player = interaction.options.getString('username');
 
-    if (userData.platform && userData.username) {
+    //already linked
+    var alreadyLinkedEmbed = new EmbedBuilder()
+        .setTitle(`${lang[langOpt].link.line_1}`)
+        .setDescription(
+            `${lang[langOpt].link.line_2}:` +
+            `\n**${userData.username}** [${userData.platform}]`)
+        .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
+        .setTimestamp()
+        .setColor("Red");
 
-        var alreadyLinkedEmbed = new EmbedBuilder()
-            .setTitle(`${lang[langOpt].link.line_1}`)
-            .setDescription(
-                `${lang[langOpt].link.line_2}:` +
-                `\n**${userData.username}** [${userData.platform}]`)
-            .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
-            .setTimestamp()
-            .setColor("Red");
+    if (userData.platform && userData.username) return interaction.editReply({ embeds: [alreadyLinkedEmbed], ephemeral: userData.invisible });
 
-        return interaction.editReply({ embeds: [alreadyLinkedEmbed], ephemeral: userData.invisible });
-    } else {
-        var url = encodeURI(`https://api.mozambiquehe.re/bridge?version=5&platform=${platform}&player=${player}&auth=${auth}`);
-        fetch(url)
-            .then(res => {
-                if (res.status === 200) { return res.json() } else {
-                    handleError(interaction, userData, res.status)
-                    return Promise.reject('Error occurred');
-                }
-            })
-            .then(async data => {
-                if (!data || !data?.global || !data?.global?.name || data?.global?.name === '') return sentLookUpError(interaction, userData);
+    var url = encodeURI(`https://api.mozambiquehe.re/bridge?version=5&platform=${platform}&player=${player}&auth=${auth}`);
+    fetch(url)
+        .then(res => res.status === 200 ? res.json() : handleError(interaction, userData, res.status))
+        .then(async data => {
+            checkData(data, interaction, userData);
 
-                await setDoc(doc(db, 'users', interaction.user.id), {
-                    platform: platform,
-                    username: data.global.name
-                }, { merge: true });
+            //succesfully linked
+            var linkedEmbed = new EmbedBuilder()
+                .setTitle(`${lang[langOpt].link.line_3}`)
+                .setDescription(`${lang[langOpt].link.line_4}:` +
+                    `\n**${data.global.name}** [${platform}]`)
+                .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
+                .setTimestamp()
+                .setColor("Green");
 
-                var linkedEmbed = new EmbedBuilder()
-                    .setTitle(`${lang[langOpt].link.line_3}`)
-                    .setDescription(`${lang[langOpt].link.line_4}:` +
-                        `\n**${data.global.name}** [${platform}]`)
-                    .setFooter({ text: `${interaction.client.user.username} ❤️`, iconURL: interaction.client.user.displayAvatarURL() })
-                    .setTimestamp()
-                    .setColor("Green");
+            await setDoc(doc(db, 'users', interaction.user.id), {
+                platform: platform,
+                username: data.global.name
+            }, { merge: true });
 
-                interaction.editReply({ embeds: [linkedEmbed], ephemeral: userData.invisible });
-            }).catch(error => { console.error('Fetch error:', error) })
-    }
+            return interaction.editReply({ embeds: [linkedEmbed], ephemeral: userData.invisible });
+        }).catch(error => { console.error('Fetch error:', error) });
 }
 
 module.exports = {