+++ src/utilities/timedEvents/botStatus.js
@@ -7,14 +7,18 @@ module.exports = {
 
         await sleep(5000);
 
+        //check for old activity
         let oldActivity = client.user.presence.activities[0];
         if (!oldActivity) return setMapData(client, auth);
 
+        //check for [] in activity
         let match = oldActivity.name.match(/\[(.*?)\]/);
-        if (!match) return setMapData(client, auth); //no []
+        if (!match) return setMapData(client, auth);
 
+        //check for text in the []
         let activityStr = match[1].match(/\d+/);
-        if (!activityStr) return setMapData(); // no text in []
+        if (!activityStr) return setMapData();
+
 
         let prefix = match[1].replace(activityStr[0], "").trim();
         if (!prefix === "min") return setMapData(client, auth);